Aggregate cells

Combine cells into groups based on shared variables and aggregate feature counts.

# S4 method for Seurat
aggregate_cells(
  .data,
  .sample = NULL,
  slot = "data",
  assays = NULL,
  aggregation_function = Matrix::rowSums
)

Arguments

.data

A tidySingleCellExperiment object

.sample

A vector of variables by which cells are aggregated

slot

The slot to which the function is applied

assays

The assay to which the function is applied

aggregation_function

The method of cell-feature value aggregation

Value

A tibble object

An object containing the information.for the specified features

Examples

data("pbmc_small")
pbmc_small |>
  aggregate_cells(c(groups, letter.idents), assays = "RNA")
#> Joining with `by = join_by(letter.idents, groups)`
#> # A tibble: 920 × 8
#>    .feature  .sample   RNA letter.idents groups .aggregated_cells orig.ident   
#>    <chr>     <chr>   <dbl> <fct>         <chr>              <int> <fct>        
#>  1 MS4A1     g2___A  35.5  A             g2                    23 SeuratProject
#>  2 CD79B     g2___A  39.3  A             g2                    23 SeuratProject
#>  3 CD79A     g2___A  29.5  A             g2                    23 SeuratProject
#>  4 HLA-DRA   g2___A  73.9  A             g2                    23 SeuratProject
#>  5 TCL1A     g2___A  29.5  A             g2                    23 SeuratProject
#>  6 HLA-DQB1  g2___A  37.6  A             g2                    23 SeuratProject
#>  7 HVCN1     g2___A  22.4  A             g2                    23 SeuratProject
#>  8 HLA-DMB   g2___A  24.7  A             g2                    23 SeuratProject
#>  9 LTB       g2___A  91.9  A             g2                    23 SeuratProject
#> 10 LINC00926 g2___A   9.91 A             g2                    23 SeuratProject
#> # ℹ 910 more rows
#> # ℹ 1 more variable: RNA_snn_res.0.8 <fct>