Extract and join information for features.

join_features() extracts and joins information for specified features

# S4 method for Seurat
join_features(
  .data,
  features = NULL,
  all = FALSE,
  exclude_zeros = FALSE,
  shape = "long",
  ...
)

Arguments

.data

A Seurat object

features

A vector of feature identifiers to join

all

If TRUE return all

exclude_zeros

If TRUE exclude zero values

shape

Format of the returned table "long" or "wide"

...

Parameters to pass to join wide, i.e. assay name to extract feature abundance from and gene prefix, for shape="wide"

Value

An object containing the information.for the specified features

An object containing the information.for the specified features

Details

This function extracts information for specified features and returns the information in either long or wide format.

Examples

data("pbmc_small")
pbmc_small %>% 
join_features(features = c("HLA-DRA", "LYZ"))
#> tidyseurat says: This operation lead to duplicated cell names. A data frame is returned for independent data analysis.
#> # A tibble: 160 × 31
#>    .cell  .feat…¹ .abun…² orig.…³ nCoun…⁴ nFeat…⁵ RNA_s…⁶ lette…⁷ groups RNA_s…⁸
#>    <chr>  <chr>     <dbl> <fct>     <dbl>   <int> <fct>   <fct>   <chr>  <fct>  
#>  1 ATGCC… HLA-DRA    0    Seurat…      70      47 0       A       g2     0      
#>  2 ATGCC… LYZ        4.97 Seurat…      70      47 0       A       g2     0      
#>  3 CATGG… HLA-DRA    4.78 Seurat…      85      52 0       A       g1     0      
#>  4 CATGG… LYZ        4.78 Seurat…      85      52 0       A       g1     0      
#>  5 GAACC… HLA-DRA    0    Seurat…      87      50 1       B       g2     0      
#>  6 GAACC… LYZ        4.75 Seurat…      87      50 1       B       g2     0      
#>  7 TGACT… HLA-DRA    0    Seurat…     127      56 0       A       g2     0      
#>  8 TGACT… LYZ        0    Seurat…     127      56 0       A       g2     0      
#>  9 AGTCA… HLA-DRA    4.07 Seurat…     173      53 0       A       g2     0      
#> 10 AGTCA… LYZ        0    Seurat…     173      53 0       A       g2     0      
#> # … with 150 more rows, 21 more variables: PC_1 <dbl>, PC_2 <dbl>, PC_3 <dbl>,
#> #   PC_4 <dbl>, PC_5 <dbl>, PC_6 <dbl>, PC_7 <dbl>, PC_8 <dbl>, PC_9 <dbl>,
#> #   PC_10 <dbl>, PC_11 <dbl>, PC_12 <dbl>, PC_13 <dbl>, PC_14 <dbl>,
#> #   PC_15 <dbl>, PC_16 <dbl>, PC_17 <dbl>, PC_18 <dbl>, PC_19 <dbl>,
#> #   tSNE_1 <dbl>, tSNE_2 <dbl>, and abbreviated variable names ¹​.feature,
#> #   ²​.abundance_RNA, ³​orig.ident, ⁴​nCount_RNA, ⁵​nFeature_RNA, ⁶​RNA_snn_res.0.8,
#> #   ⁷​letter.idents, ⁸​RNA_snn_res.1