Subset rows using their positions

`slice()` lets you index rows by their (integer) locations. It allows you to select, remove, and duplicate rows. It is accompanied by a number of helpers for common use cases:

* `slice_head()` and `slice_tail()` select the first or last rows. * `slice_sample()` randomly selects rows. * `slice_min()` and `slice_max()` select rows with highest or lowest values of a variable.

If `.data` is a [grouped_df], the operation will be performed on each group, so that (e.g.) `slice_head(df, n = 5)` will select the first five rows in each group.

slice(.data, ..., .preserve = FALSE)

Arguments

.data
A data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr). See *Methods*, below, for more details.

...
For `slice()`: <[`data-masking`][dplyr_data_masking]> Integer row values.

Provide either positive values to keep, or negative values to drop. The values provided must be either all positive or all negative. Indices beyond the number of rows in the input are silently ignored.

For `slice_helpers()`, these arguments are passed on to methods.

If `n` is greater than the number of rows in the group (or `prop > 1`), the result will be silently truncated to the group size. If the `prop`ortion of a group size is not an integer, it is rounded down.

.preserve
when `FALSE` (the default), the grouping structure is recalculated based on the resulting data, otherwise it is kept as is.

Value

An object of the same type as `.data`. The output has the following properties:

* Each row may appear 0, 1, or many times in the output. * Columns are not modified. * Groups are not modified. * Data frame attributes are preserved.

Details

Slice does not work with relational databases because they have no intrinsic notion of row order. If you want to perform the equivalent operation, use [filter()] and [row_number()].

Methods

These function are **generic**s, which means that packages can provide implementations (methods) for other classes. See the documentation of individual methods for extra arguments and differences in behaviour.

Methods available in currently loaded packages:

* `slice()`: base (index), dplyr (data.frame, default) . * `slice_head()`: dplyr (data.frame) . * `slice_tail()`: dplyr (data.frame) . * `slice_min()`: dplyr (data.frame) . * `slice_max()`: dplyr (data.frame) . * `slice_sample()`: dplyr (data.frame) .

See also

Examples

`%>%` = magrittr::`%>%`
pbmc_small %>% tidy %>% slice(1)

#> Warning: `slice_()` is deprecated as of dplyr 0.7.0.
#> Please use `slice()` instead.
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_warnings()` to see where this warning was generated.

#> Error in UseMethod("slice_"): no applicable method for 'slice_' applied to an object of class "c('tidyseurat', 'Seurat')"