Coerce lists, matrices, and more to data frames

`r lifecycle::badge("maturing")`

`as_tibble()` turns an existing object, such as a data frame or matrix, into a so-called tibble, a data frame with class [`tbl_df`]. This is in contrast with [tibble()], which builds a tibble from individual columns. `as_tibble()` is to [`tibble()`] as [base::as.data.frame()] is to [base::data.frame()].

`as_tibble()` is an S3 generic, with methods for: * [`data.frame`][base::data.frame()]: Thin wrapper around the `list` method that implements tibble's treatment of [rownames]. * [`matrix`][methods::matrix-class], [`poly`][stats::poly()], [`ts`][stats::ts()], [`table`][base::table()] * Default: Other inputs are first coerced with [base::as.data.frame()].

`r lifecycle::badge("maturing")`

`glimpse()` is like a transposed version of `print()`: columns run down the page, and data runs across. This makes it possible to see every column in a data frame. It's a little like [str()] applied to a data frame but it tries to show you as much data as possible. (And it always shows the underlying data, even when applied to a remote data source.)

This generic will be moved to pillar, and reexported from there as soon as it becomes available.

Arguments

rownames
How to treat existing row names of a data frame or matrix: * `NULL`: remove row names. This is the default. * `NA`: keep row names. * A string: the name of a new column. Existing rownames are transferred into this column and the `row.names` attribute is deleted. Read more in [rownames].

.name_repair
see tidyr

For compatibility only, do not use for new code.

x
An object to glimpse at.

width
Width of output: defaults to the setting of the option `tibble.width` (if finite) or the width of the console.

...
Unused, for extensibility.

Value

A tibble

x original x is (invisibly) returned, allowing `glimpse()` to be used within a data pipe line.

Row names

The default behavior is to silently remove row names.

New code should explicitly convert row names to a new column using the `rownames` argument.

For existing code that relies on the retention of row names, call `pkgconfig::set_config("tibble::rownames" = NA)` in your script or in your package's [.onLoad()] function.

Life cycle

Using `as_tibble()` for vectors is superseded as of version 3.0.0, prefer the more expressive maturing `as_tibble_row()` and `as_tibble_col()` variants for new code.

S3 methods

`glimpse` is an S3 generic with a customised method for `tbl`s and `data.frames`, and a default method that calls [str()].

See also

Examples

pbmc_small %>% tidy %>% as_tibble()

#> # A tibble: 80 x 30
#>    cell  orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.8 letter.idents groups
#>    <chr> <fct>           <dbl>        <int> <fct>           <fct>         <chr> 
#>  1 ATGC… SeuratPro…         70           47 0               A             g2    
#>  2 CATG… SeuratPro…         85           52 0               A             g1    
#>  3 GAAC… SeuratPro…         87           50 1               B             g2    
#>  4 TGAC… SeuratPro…        127           56 0               A             g2    
#>  5 AGTC… SeuratPro…        173           53 0               A             g2    
#>  6 TCTG… SeuratPro…         70           48 0               A             g1    
#>  7 TGGT… SeuratPro…         64           36 0               A             g1    
#>  8 GCAG… SeuratPro…         72           45 0               A             g1    
#>  9 GATA… SeuratPro…         52           36 0               A             g1    
#> 10 AATG… SeuratPro…        100           41 0               A             g1    
#> # … with 70 more rows, and 23 more variables: RNA_snn_res.1 <fct>, file <chr>,
#> #   PC_1 <dbl>, PC_2 <dbl>, PC_3 <dbl>, PC_4 <dbl>, PC_5 <dbl>, PC_6 <dbl>,
#> #   PC_7 <dbl>, PC_8 <dbl>, PC_9 <dbl>, PC_10 <dbl>, PC_11 <dbl>, PC_12 <dbl>,
#> #   PC_13 <dbl>, PC_14 <dbl>, PC_15 <dbl>, PC_16 <dbl>, PC_17 <dbl>,
#> #   PC_18 <dbl>, PC_19 <dbl>, tSNE_1 <dbl>, tSNE_2 <dbl>

pbmc_small %>% tidy %>% glimpse()

#> Rows: 80
#> Columns: 30
#> $ cell            <chr> "ATGCCAGAACGACT", "CATGGCCTGTGCAT", "GAACCTGATGAACC", …
#> $ orig.ident      <fct> SeuratProject, SeuratProject, SeuratProject, SeuratPro…
#> $ nCount_RNA      <dbl> 70, 85, 87, 127, 173, 70, 64, 72, 52, 100, 62, 191, 10…
#> $ nFeature_RNA    <int> 47, 52, 50, 56, 53, 48, 36, 45, 36, 41, 31, 61, 41, 26…
#> $ RNA_snn_res.0.8 <fct> 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ letter.idents   <fct> A, A, B, A, A, A, A, A, A, A, A, A, A, A, A, A, A, A, …
#> $ groups          <chr> "g2", "g1", "g2", "g2", "g2", "g1", "g1", "g1", "g1", …
#> $ RNA_snn_res.1   <fct> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 2, 2, 2, 2, 2, 2, 2, 2, …
#> $ file            <chr> "../data/sample2/outs/filtered_feature_bc_matrix/", ".…
#> $ PC_1            <dbl> -0.77403708, -0.02602702, -0.45650250, -0.81163243, -0…
#> $ PC_2            <dbl> -0.89964608, -0.34667947, 0.17958114, -1.37953401, -0.…
#> $ PC_3            <dbl> -0.2493078, 0.6651668, 1.3175907, -1.0019320, -0.24930…
#> $ PC_4            <dbl> 0.55859480, 0.41828998, 2.01372104, 0.13905026, 0.5585…
#> $ PC_5            <dbl> 0.46508378, 0.58532045, -0.48188513, -1.59822320, 0.46…
#> $ PC_6            <dbl> -1.38727138, -1.17380941, 0.35757129, -0.23157629, -1.…
#> $ PC_7            <dbl> -0.1277598, 0.1195137, -1.3821372, 0.1396306, -0.12775…
#> $ PC_8            <dbl> -0.29002273, -0.21363844, -1.72040130, 0.72629941, -0.…
#> $ PC_9            <dbl> 0.37365607, 0.34941666, 0.81703026, 0.67584015, 0.3736…
#> $ PC_10           <dbl> 0.44682478, 1.20616303, 0.06372349, -0.73786224, 0.446…
#> $ PC_11           <dbl> -0.24265143, 0.90313003, 0.08629753, 0.79933396, -0.24…
#> $ PC_12           <dbl> -0.16502284, -0.39257588, -0.46815637, -0.61207526, -0…
#> $ PC_13           <dbl> -0.013565041, 0.803542544, -0.128876042, 0.776509775, …
#> $ PC_14           <dbl> -0.42514591, -1.03526716, -0.86425209, 0.34403811, -0.…
#> $ PC_15           <dbl> -0.14646171, 0.55604780, 0.72772498, -0.48998525, -0.1…
#> $ PC_16           <dbl> -0.394102365, 1.520806392, -0.070302297, -0.303268701,…
#> $ PC_17           <dbl> 7.561356e-02, 1.995033e-01, -6.991638e-01, 1.306050e-0…
#> $ PC_18           <dbl> 0.019799375, -0.880980524, 0.900604550, -0.213501542, …
#> $ PC_19           <dbl> 0.04491314, 0.19911188, -0.32448103, 0.06828601, 0.044…
#> $ tSNE_1          <dbl> 0.8675977, -7.3925306, -28.2064258, 16.3480689, 1.9113…
#> $ tSNE_2          <dbl> -8.1007483, -8.7717451, 0.2410102, -11.1633255, -11.19…