unnest

unnest

nest

unnest(
  .data,
  cols,
  ...,
  keep_empty = FALSE,
  ptype = NULL,
  names_sep = NULL,
  names_repair = "check_unique"
)

# S3 method for default
unnest(
  .data,
  cols,
  ...,
  keep_empty = FALSE,
  ptype = NULL,
  names_sep = NULL,
  names_repair = "check_unique"
)

# S3 method for tidyseurat_nested
unnest(
  .data,
  cols,
  ...,
  keep_empty = FALSE,
  ptype = NULL,
  names_sep = NULL,
  names_repair = "check_unique"
)

nest(.data, ...)

# S3 method for default
nest(.data, ...)

# S3 method for tidyseurat
nest(.data, ...)

Arguments

.data
A tbl. (See tidyr)

cols
<[`tidy-select`][tidyr_tidy_select]> Columns to unnest. If you `unnest()` multiple columns, parallel entries must be of compatible sizes, i.e. they're either equal or length 1 (following the standard tidyverse recycling rules).

...
Name-variable pairs of the form new_col = c(col1, col2, col3) (See tidyr)

keep_empty
See tidyr::unnest

ptype
See tidyr::unnest

names_sep
If `NULL`, the default, the names will be left as is. In `nest()`, inner names will come from the former outer names; in `unnest()`, the new outer names will come from the inner names.

If a string, the inner and outer names will be used together. In `nest()`, the names of the new outer columns will be formed by pasting together the outer and the inner column names, separated by `names_sep`. In `unnest()`, the new inner names will have the outer names (+ `names_sep`) automatically stripped. This makes `names_sep` roughly symmetric between nesting and unnesting.

names_repair
See tidyr::unnest

Value

A tidyseurat objector a tibble depending on input

A tidyseurat objector a tibble depending on input

Examples

library(dplyr)
pbmc_small %>% tidy %>% nest(data = -groups) %>% unnest(data) 

#> # A tibble: 80 x 8
#>    cell  orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.8 letter.idents
#>    <chr> <chr>           <dbl>        <int> <chr>           <chr>        
#>  1 ATGC… SeuratPro…         70           47 0               A            
#>  2 GAAC… SeuratPro…         87           50 1               B            
#>  3 TGAC… SeuratPro…        127           56 0               A            
#>  4 AGTC… SeuratPro…        173           53 0               A            
#>  5 AGGT… SeuratPro…         62           31 0               A            
#>  6 GGGT… SeuratPro…        101           41 0               A            
#>  7 CATG… SeuratPro…         51           26 0               A            
#>  8 TACG… SeuratPro…         99           45 0               A            
#>  9 GTAA… SeuratPro…         67           33 0               A            
#> 10 TACA… SeuratPro…        109           41 0               A            
#> # … with 70 more rows, and 2 more variables: RNA_snn_res.1 <chr>, groups <chr>

library(dplyr)
pbmc_small %>% tidy %>% nest(data = -groups) %>% unnest(data) 

#> # A tibble: 80 x 8
#>    cell  orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.8 letter.idents
#>    <chr> <chr>           <dbl>        <int> <chr>           <chr>        
#>  1 ATGC… SeuratPro…         70           47 0               A            
#>  2 GAAC… SeuratPro…         87           50 1               B            
#>  3 TGAC… SeuratPro…        127           56 0               A            
#>  4 AGTC… SeuratPro…        173           53 0               A            
#>  5 AGGT… SeuratPro…         62           31 0               A            
#>  6 GGGT… SeuratPro…        101           41 0               A            
#>  7 CATG… SeuratPro…         51           26 0               A            
#>  8 TACG… SeuratPro…         99           45 0               A            
#>  9 GTAA… SeuratPro…         67           33 0               A            
#> 10 TACA… SeuratPro…        109           41 0               A            
#> # … with 70 more rows, and 2 more variables: RNA_snn_res.1 <chr>, groups <chr>