unnest_seurat

unnest_seurat

unnest_seurat(
  data,
  cols,
  ...,
  keep_empty = FALSE,
  ptype = NULL,
  names_sep = NULL,
  names_repair = "check_unique",
  .drop,
  .id,
  .sep,
  .preserve
)

Arguments

data

A tbl. (See tidyr)

cols

<[`tidy-select`][tidyr_tidy_select]> Columns to unnest. If you `unnest()` multiple columns, parallel entries must be of compatible sizes, i.e. they're either equal or length 1 (following the standard tidyverse recycling rules).

...

<[`tidy-select`][tidyr_tidy_select]> Columns to nest, specified using name-variable pairs of the form `new_col=c(col1, col2, col3)`. The right hand side can be any valid tidy select expression.

keep_empty

See tidyr::unnest

ptype

See tidyr::unnest

names_sep

If `NULL`, the default, the names will be left as is. In `nest()`, inner names will come from the former outer names; in `unnest()`, the new outer names will come from the inner names.

If a string, the inner and outer names will be used together. In `nest()`, the names of the new outer columns will be formed by pasting together the outer and the inner column names, separated by `names_sep`. In `unnest()`, the new inner names will have the outer names (+ `names_sep`) automatically stripped. This makes `names_sep` roughly symmetric between nesting and unnesting.

names_repair

See tidyr::unnest

.drop

See tidyr::unnest

.id

tidyr::unnest

.sep

tidyr::unnest

.preserve

See tidyr::unnest

Value

A tidySingleCellExperiment objector a tibble depending on input

Examples

library(dplyr)
pbmc_small %>%

    nest(data=-groups) %>%
    unnest_seurat(data)
#> # A Seurat-tibble abstraction: 80 × 8
#> # Features=230 | Cells=80 | Active assay=RNA | Assays=RNA
#>    .cell        orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.8 letter.idents
#>    <chr>        <chr>           <dbl>        <int> <chr>           <chr>        
#>  1 ATGCCAGAACG… SeuratPro…         70           47 0               A            
#>  2 GAACCTGATGA… SeuratPro…         87           50 1               B            
#>  3 TGACTGGATTC… SeuratPro…        127           56 0               A            
#>  4 AGTCAGACTGC… SeuratPro…        173           53 0               A            
#>  5 AGGTCATGAGT… SeuratPro…         62           31 0               A            
#>  6 GGGTAACTCTA… SeuratPro…        101           41 0               A            
#>  7 CATGAGACACG… SeuratPro…         51           26 0               A            
#>  8 TACGCCACTCC… SeuratPro…         99           45 0               A            
#>  9 GTAAGCACTCA… SeuratPro…         67           33 0               A            
#> 10 TACATCACGCT… SeuratPro…        109           41 0               A            
#> # ℹ 70 more rows
#> # ℹ 2 more variables: RNA_snn_res.1 <chr>, groups <chr>