Coerce lists, matrices, and more to data frames

`r lifecycle::badge("maturing")`

`as_tibble()` turns an existing object, such as a data frame or matrix, into a so-called tibble, a data frame with class [`tbl_df`]. This is in contrast with [tibble()], which builds a tibble from individual columns. `as_tibble()` is to [`tibble()`] as [base::as.data.frame()] is to [base::data.frame()].

`as_tibble()` is an S3 generic, with methods for: * [`data.frame`][base::data.frame()]: Thin wrapper around the `list` method that implements tibble's treatment of [rownames]. * [`matrix`][methods::matrix-class], [`poly`][stats::poly()], [`ts`][stats::ts()], [`table`][base::table()] * Default: Other inputs are first coerced with [base::as.data.frame()].

as_tibble(
  x,
  ...,
  .name_repair = c("check_unique", "unique", "universal", "minimal"),
  rownames = pkgconfig::get_config("tibble::rownames", NULL)
)

Arguments

x
A data frame, list, matrix, or other object that could reasonably be coerced to a tibble.

...
Unused, for extensibility.

.name_repair
see tidyr

For compatibility only, do not use for new code.

rownames
How to treat existing row names of a data frame or matrix: * `NULL`: remove row names. This is the default. * `NA`: keep row names. * A string: the name of a new column. Existing rownames are transferred into this column and the `row.names` attribute is deleted. Read more in [rownames].

Value

A tibble

Row names

The default behavior is to silently remove row names.

New code should explicitly convert row names to a new column using the `rownames` argument.

For existing code that relies on the retention of row names, call `pkgconfig::set_config("tibble::rownames" = NA)` in your script or in your package's [.onLoad()] function.

Life cycle

Using `as_tibble()` for vectors is superseded as of version 3.0.0, prefer the more expressive maturing `as_tibble_row()` and `as_tibble_col()` variants for new code.

See also

Examples

pbmc_small %>% tidy %>% as_tibble()

#> # A tibble: 80 x 16
#>    cell  orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.8 letter.idents groups
#>    <chr> <fct>           <dbl>        <int> <fct>           <fct>         <chr> 
#>  1 ATGC… SeuratPro…         70           47 0               A             g2    
#>  2 CATG… SeuratPro…         85           52 0               A             g1    
#>  3 GAAC… SeuratPro…         87           50 1               B             g2    
#>  4 TGAC… SeuratPro…        127           56 0               A             g2    
#>  5 AGTC… SeuratPro…        173           53 0               A             g2    
#>  6 TCTG… SeuratPro…         70           48 0               A             g1    
#>  7 TGGT… SeuratPro…         64           36 0               A             g1    
#>  8 GCAG… SeuratPro…         72           45 0               A             g1    
#>  9 GATA… SeuratPro…         52           36 0               A             g1    
#> 10 AATG… SeuratPro…        100           41 0               A             g1    
#> # … with 70 more rows, and 9 more variables: RNA_snn_res.1 <fct>, file <chr>,
#> #   PC_1 <dbl>, PC_2 <dbl>, PC_3 <dbl>, PC_4 <dbl>, PC_5 <dbl>, tSNE_1 <dbl>,
#> #   tSNE_2 <dbl>