Efficiently bind multiple data frames by row and column

This is an efficient implementation of the common pattern of `do.call(rbind, dfs)` or `do.call(cbind, dfs)` for binding many data frames into one.

Arguments

...

Data frames to combine.

Each argument can either be a data frame, a list that could be a data frame, or a list of data frames.

When row-binding, columns are matched by name, and any missing columns will be filled with NA.

When column-binding, rows are matched by position, so all data frames must have the same number of rows. To match by value, not position, see mutate-joins.

.id

Data frame identifier.

When `.id` is supplied, a new column of identifiers is created to link each row to its original data frame. The labels are taken from the named arguments to `bind_rows()`. When a list of data frames is supplied, the labels are taken from the names of the list. If no names are found a numeric sequence is used instead.

add.cell.ids

from Seurat 3.0 A character vector of length(x = c(x, y)). Appends the corresponding values to the start of each objects' cell names.

Value

`bind_rows()` and `bind_cols()` return the same type as the first input, either a data frame, `tbl_df`, or `grouped_df`.

Details

The output of `bind_rows()` will contain a column if that column appears in any of the inputs.

Examples

`%>%` = magrittr::`%>%`
tt = pbmc_small
bind_rows(    tt, tt  )
#> Warning: Some cell names are duplicated across objects provided. Renaming to enforce unique cell names.
#> # A Seurat-tibble abstraction: 160 × 8
#> # Features=230 | Cells=160 | Active assay=RNA | Assays=RNA
#>    .cell            orig.ident    nCoun…¹ nFeat…² RNA_s…³ lette…⁴ groups RNA_s…⁵
#>    <chr>            <chr>           <dbl>   <int> <chr>   <chr>   <chr>  <chr>  
#>  1 ATGCCAGAACGACT_1 SeuratProject      70      47 0       A       g2     0      
#>  2 CATGGCCTGTGCAT_1 SeuratProject      85      52 0       A       g1     0      
#>  3 GAACCTGATGAACC_1 SeuratProject      87      50 1       B       g2     0      
#>  4 TGACTGGATTCTCA_1 SeuratProject     127      56 0       A       g2     0      
#>  5 AGTCAGACTGCACA_1 SeuratProject     173      53 0       A       g2     0      
#>  6 TCTGATACACGTGT_1 SeuratProject      70      48 0       A       g1     0      
#>  7 TGGTATCTAAACAG_1 SeuratProject      64      36 0       A       g1     0      
#>  8 GCAGCTCTGTTTCT_1 SeuratProject      72      45 0       A       g1     0      
#>  9 GATATAACACGCAT_1 SeuratProject      52      36 0       A       g1     0      
#> 10 AATGTTGACAGTCA_1 SeuratProject     100      41 0       A       g1     0      
#> # … with 150 more rows, and abbreviated variable names ¹​nCount_RNA,
#> #   ²​nFeature_RNA, ³​RNA_snn_res.0.8, ⁴​letter.idents, ⁵​RNA_snn_res.1

tt_bind = tt %>% select(nCount_RNA ,nFeature_RNA)
#> tidyseurat says: Key columns are missing. A data frame is returned for independent data analysis.
tt %>% bind_cols(tt_bind)
#> New names:
#> • `nCount_RNA` -> `nCount_RNA...2`
#> • `nFeature_RNA` -> `nFeature_RNA...3`
#> • `nCount_RNA` -> `nCount_RNA...8`
#> • `nFeature_RNA` -> `nFeature_RNA...9`
#> # A Seurat-tibble abstraction: 80 × 17
#> # Features=230 | Cells=80 | Active assay=RNA | Assays=RNA
#>    .cell  orig.…¹ nCoun…² nFeat…³ RNA_s…⁴ lette…⁵ groups RNA_s…⁶ nCoun…⁷ nFeat…⁸
#>    <chr>  <fct>     <dbl>   <int> <fct>   <fct>   <chr>  <fct>     <dbl>   <int>
#>  1 ATGCC… Seurat…      70      47 0       A       g2     0            70      47
#>  2 CATGG… Seurat…      85      52 0       A       g1     0            85      52
#>  3 GAACC… Seurat…      87      50 1       B       g2     0            87      50
#>  4 TGACT… Seurat…     127      56 0       A       g2     0           127      56
#>  5 AGTCA… Seurat…     173      53 0       A       g2     0           173      53
#>  6 TCTGA… Seurat…      70      48 0       A       g1     0            70      48
#>  7 TGGTA… Seurat…      64      36 0       A       g1     0            64      36
#>  8 GCAGC… Seurat…      72      45 0       A       g1     0            72      45
#>  9 GATAT… Seurat…      52      36 0       A       g1     0            52      36
#> 10 AATGT… Seurat…     100      41 0       A       g1     0           100      41
#> # … with 70 more rows, 7 more variables: PC_1 <dbl>, PC_2 <dbl>, PC_3 <dbl>,
#> #   PC_4 <dbl>, PC_5 <dbl>, tSNE_1 <dbl>, tSNE_2 <dbl>, and abbreviated
#> #   variable names ¹​orig.ident, ²​nCount_RNA...2, ³​nFeature_RNA...3,
#> #   ⁴​RNA_snn_res.0.8, ⁵​letter.idents, ⁶​RNA_snn_res.1, ⁷​nCount_RNA...8,
#> #   ⁸​nFeature_RNA...9