Arrange rows by column values

`arrange()` order the rows of a data frame rows by the values of selected columns.

Unlike other dplyr verbs, `arrange()` largely ignores grouping; you need to explicit mention grouping variables (or use `by_group = TRUE`) in order to group by them, and functions of variables are evaluated once per data frame, not once per group.

`filter()` retains the rows where the conditions you provide a `TRUE`. Note that, unlike base subsetting with `[`, rows where the condition evaluates to `NA` are dropped.

Most data operations are done on groups defined by variables. `group_by()` takes an existing tbl and converts it into a grouped tbl where operations are performed "by group". `ungroup()` removes grouping.

`summarise()` creates a new data frame. It will have one (or more) rows for each combination of grouping variables; if there are no grouping variables, the output will have a single row summarising all observations in the input. It will contain one column for each grouping variable and one column for each of the summary statistics that you have specified.

`summarise()` and `summarize()` are synonyms.

`mutate()` adds new variables and preserves existing ones; `transmute()` adds new variables and drops existing ones. New variables overwrite existing variables of the same name. Variables can be removed by setting their value to `NULL`.

Rename individual variables using `new_name = old_name` syntax.

`rowwise()` is used for the results of [do()] when you create list-variables. It is also useful to support arbitrary complex operations that need to be applied to each row.

`slice()` lets you index rows by their (integer) locations. It allows you to select, remove, and duplicate rows. It is accompanied by a number of helpers for common use cases:

Select (and optionally rename) variables in a data frame, using a concise mini-language that makes it easy to refer to variables based on their name (e.g. `a:f` selects all columns from `a` on the left to `f` on the right). You can also use predicate functions like is.numeric to select variables based on their properties.

Sample n rows from a table

`pull()` is similar to `$`. It's mostly useful because it looks a little nicer in pipes, it also works with remote data frames, and it can optionally name the output.

bind_rows(..., .id = NULL, add.cell.ids = NULL)

bind_cols(..., .id = NULL)

Arguments

...
For use by methods.

.id
Data frame identifier.

When `.id` is supplied, a new column of identifiers is created to link each row to its original data frame. The labels are taken from the named arguments to `bind_rows()`. When a list of data frames is supplied, the labels are taken from the names of the list. If no names are found a numeric sequence is used instead.

add.cell.ids
from Seurat 3.0 A character vector of length(x = c(x, y)). Appends the corresponding values to the start of each objects' cell names.

.by_group
If TRUE, will sort first by grouping variable. Applies to grouped data frames only.

.keep_all
If TRUE, keep all variables in .data. If a combination of ... is not distinct, this keeps the first row of values. (See dplyr)

.preserve
when `FALSE` (the default), the grouping structure is recalculated based on the resulting data, otherwise it is kept as is.

.add
When `FALSE`, the default, `group_by()` will override existing groups. To add to the existing groups, use `.add = TRUE`.

This argument was previously called `add`, but that prevented creating a new grouping variable called `add`, and conflicts with our naming conventions.

.drop
When `.drop = TRUE`, empty groups are dropped. See group_by_drop_default() for what the default value is for this argument.

.data
Input data frame.

x
tbls to join. (See dplyr)

y
tbls to join. (See dplyr)

by
A character vector of variables to join by. (See dplyr)

copy
If x and y are not from the same data source, and copy is TRUE, then y will be copied into the same src as x. (See dplyr)

suffix
If there are non-joined duplicate variables in x and y, these suffixes will be added to the output to disambiguate them. Should be a character vector of length 2. (See dplyr)

tbl
A data.frame.

size
<[`tidy-select`][dplyr_select]> For `sample_n()`, the number of rows to select. For `sample_frac()`, the fraction of rows to select. If `tbl` is grouped, `size` applies to each group.

replace
Sample with or without replacement?

weight
<[`tidy-select`][dplyr_select]> Sampling weights. This must evaluate to a vector of non-negative numbers the same length as the input. Weights are automatically standardised to sum to 1.

.env
DEPRECATED.

name
An optional parameter that specifies the column to be used as names for a named vector. Specified in a similar manner as var.

Value

An object of the same type as `.data`.

* All rows appear in the output, but (usually) in a different place. * Columns are not modified. * Groups are not modified. * Data frame attributes are preserved.

A Seurat object

An object of the same type as `.data`.

* Rows are a subset of the input, but appear in the same order. * Columns are not modified. * The number of groups may be reduced (if `.preserve` is not `TRUE`). * Data frame attributes are preserved.

A grouped data frame, unless the combination of `...` and `add` yields a non empty set of grouping columns, a regular (ungrouped) data frame otherwise.

A tibble

An object of the same type as `.data`.

For `mutate()`:

* Rows are not affected. * Existing columns will be preserved unless explicitly modified. * New columns will be added to the right of existing columns. * Columns given value `NULL` will be removed * Groups will be recomputed if a grouping variable is mutated. * Data frame attributes are preserved.

For `transmute()`:

* Rows are not affected. * Apart from grouping variables, existing columns will be remove unless explicitly kept. * Column order matches order of expressions. * Groups will be recomputed if a grouping variable is mutated. * Data frame attributes are preserved.

An object of the same type as `.data`. * Rows are not affected. * Column names are changed; column order is preserved * Data frame attributes are preserved. * Groups are updated to reflect new names.

A `tbl`

A `tbl`

A Seurat object

A Seurat object

A Seurat object

A Seurat object

An object of the same type as `.data`. The output has the following properties:

* Each row may appear 0, 1, or many times in the output. * Columns are not modified. * Groups are not modified. * Data frame attributes are preserved.

An object of the same type as `.data`. The output has the following properties:

* Rows are not affected. * Output columns are a subset of input columns, potentially with a different order. Columns will be renamed if `new_name = old_name` form is used. * Data frame attributes are preserved. * Groups are maintained; you can't select off grouping variables.

A Seurat object

A vector the same size as `.data`.

Details

## Locales The sort order for character vectors will depend on the collating sequence of the locale in use: see locales().

## Missing values Unlike base sorting with `sort()`, `NA` are: * always sorted to the end for local data, even when wrapped with `desc()`. * treated differently for remote data, depending on the backend.

dplyr is not yet smart enough to optimise filtering optimisation on grouped datasets that don't need grouped calculations. For this reason, filtering is often considerably faster on ungroup()ed data.

Slice does not work with relational databases because they have no intrinsic notion of row order. If you want to perform the equivalent operation, use [filter()] and [row_number()].

Methods

This function is a **generic**, which means that packages can provide implementations (methods) for other classes. See the documentation of individual methods for extra arguments and differences in behaviour.

These function are **generic**s, which means that packages can provide implementations (methods) for other classes. See the documentation of individual methods for extra arguments and differences in behaviour.

Methods available in currently loaded packages:

These function are **generic**s, which means that packages can provide implementations (methods) for other classes. See the documentation of individual methods for extra arguments and differences in behaviour.

Methods available in currently loaded packages:

This function is a **generic**, which means that packages can provide implementations (methods) for other classes. See the documentation of individual methods for extra arguments and differences in behaviour.

The following methods are currently available in loaded packages:

Grouped tibbles

Because mutating expressions are computed within groups, they may yield different results on grouped tibbles. This will be the case as soon as an aggregating, lagging, or ranking function is involved. Compare this ungrouped mutate:

With the grouped equivalent:

The former normalises `mass` by the global average whereas the latter normalises by the averages within gender levels.

Examples

`%>%` = magrittr::`%>%`
pbmc_small %>%  arrange(nFeature_RNA)

#> # A Seurat-tibble abstraction: 80 x 15
#> # Features=230 | Active assay=RNA | Assays=RNA
#>    cell  orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.8 letter.idents groups
#>    <chr> <fct>           <dbl>        <int> <fct>           <fct>         <chr> 
#>  1 CATG… SeuratPro…         51           26 0               A             g2    
#>  2 GGCA… SeuratPro…        172           29 0               A             g1    
#>  3 AGTC… SeuratPro…        157           29 0               A             g1    
#>  4 GACG… SeuratPro…        202           30 0               A             g2    
#>  5 GGAA… SeuratPro…        150           30 0               A             g2    
#>  6 AGGT… SeuratPro…         62           31 0               A             g2    
#>  7 CTTC… SeuratPro…         41           32 0               A             g2    
#>  8 GTAA… SeuratPro…         67           33 0               A             g2    
#>  9 GTCA… SeuratPro…        210           33 0               A             g2    
#> 10 TGGT… SeuratPro…         64           36 0               A             g1    
#> # … with 70 more rows, and 8 more variables: RNA_snn_res.1 <fct>, PC_1 <dbl>,
#> #   PC_2 <dbl>, PC_3 <dbl>, PC_4 <dbl>, PC_5 <dbl>, tSNE_1 <dbl>, tSNE_2 <dbl>

`%>%` = magrittr::`%>%`
data("pbmc_small")
pbmc_small %>%  distinct(groups)

#> tidyseurat says: A data frame is returned for independent data analysis.

#> # A tibble: 2 x 1
#>   groups
#>   <chr> 
#> 1 g2    
#> 2 g1    

`%>%` = magrittr::`%>%`
data("pbmc_small")
pbmc_small %>%  filter(groups == "g1")

#> # A Seurat-tibble abstraction: 44 x 15
#> # Features=230 | Active assay=RNA | Assays=RNA
#>    cell  orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.8 letter.idents groups
#>    <chr> <fct>           <dbl>        <int> <fct>           <fct>         <chr> 
#>  1 CATG… SeuratPro…         85           52 0               A             g1    
#>  2 TCTG… SeuratPro…         70           48 0               A             g1    
#>  3 TGGT… SeuratPro…         64           36 0               A             g1    
#>  4 GCAG… SeuratPro…         72           45 0               A             g1    
#>  5 GATA… SeuratPro…         52           36 0               A             g1    
#>  6 AATG… SeuratPro…        100           41 0               A             g1    
#>  7 AGAG… SeuratPro…        191           61 0               A             g1    
#>  8 CTAA… SeuratPro…        168           44 0               A             g1    
#>  9 TTGG… SeuratPro…        135           45 0               A             g1    
#> 10 CATC… SeuratPro…         79           43 0               A             g1    
#> # … with 34 more rows, and 8 more variables: RNA_snn_res.1 <fct>, PC_1 <dbl>,
#> #   PC_2 <dbl>, PC_3 <dbl>, PC_4 <dbl>, PC_5 <dbl>, tSNE_1 <dbl>, tSNE_2 <dbl>

# Learn more in ?dplyr_eval
`%>%` = magrittr::`%>%`
data("pbmc_small")
pbmc_small %>%  group_by(groups)

#> tidyseurat says: A data frame is returned for independent data analysis.

#> # A tibble: 80 x 29
#> # Groups:   groups [2]
#>    cell  orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.8 letter.idents groups
#>    <chr> <fct>           <dbl>        <int> <fct>           <fct>         <chr> 
#>  1 ATGC… SeuratPro…         70           47 0               A             g2    
#>  2 CATG… SeuratPro…         85           52 0               A             g1    
#>  3 GAAC… SeuratPro…         87           50 1               B             g2    
#>  4 TGAC… SeuratPro…        127           56 0               A             g2    
#>  5 AGTC… SeuratPro…        173           53 0               A             g2    
#>  6 TCTG… SeuratPro…         70           48 0               A             g1    
#>  7 TGGT… SeuratPro…         64           36 0               A             g1    
#>  8 GCAG… SeuratPro…         72           45 0               A             g1    
#>  9 GATA… SeuratPro…         52           36 0               A             g1    
#> 10 AATG… SeuratPro…        100           41 0               A             g1    
#> # … with 70 more rows, and 22 more variables: RNA_snn_res.1 <fct>, PC_1 <dbl>,
#> #   PC_2 <dbl>, PC_3 <dbl>, PC_4 <dbl>, PC_5 <dbl>, PC_6 <dbl>, PC_7 <dbl>,
#> #   PC_8 <dbl>, PC_9 <dbl>, PC_10 <dbl>, PC_11 <dbl>, PC_12 <dbl>, PC_13 <dbl>,
#> #   PC_14 <dbl>, PC_15 <dbl>, PC_16 <dbl>, PC_17 <dbl>, PC_18 <dbl>,
#> #   PC_19 <dbl>, tSNE_1 <dbl>, tSNE_2 <dbl>

`%>%` = magrittr::`%>%`
data("pbmc_small")
pbmc_small %>%  summarise(mean(nCount_RNA))

#> tidyseurat says: A data frame is returned for independent data analysis.

#> # A tibble: 1 x 1
#>   `mean(nCount_RNA)`
#>                <dbl>
#> 1               245.

`%>%` = magrittr::`%>%`
data("pbmc_small")
pbmc_small %>%  mutate(nFeature_RNA = 1)

#> # A Seurat-tibble abstraction: 80 x 15
#> # Features=230 | Active assay=RNA | Assays=RNA
#>    cell  orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.8 letter.idents groups
#>    <chr> <fct>           <dbl>        <dbl> <fct>           <fct>         <chr> 
#>  1 ATGC… SeuratPro…         70            1 0               A             g2    
#>  2 CATG… SeuratPro…         85            1 0               A             g1    
#>  3 GAAC… SeuratPro…         87            1 1               B             g2    
#>  4 TGAC… SeuratPro…        127            1 0               A             g2    
#>  5 AGTC… SeuratPro…        173            1 0               A             g2    
#>  6 TCTG… SeuratPro…         70            1 0               A             g1    
#>  7 TGGT… SeuratPro…         64            1 0               A             g1    
#>  8 GCAG… SeuratPro…         72            1 0               A             g1    
#>  9 GATA… SeuratPro…         52            1 0               A             g1    
#> 10 AATG… SeuratPro…        100            1 0               A             g1    
#> # … with 70 more rows, and 8 more variables: RNA_snn_res.1 <fct>, PC_1 <dbl>,
#> #   PC_2 <dbl>, PC_3 <dbl>, PC_4 <dbl>, PC_5 <dbl>, tSNE_1 <dbl>, tSNE_2 <dbl>

`%>%` = magrittr::`%>%`
data("pbmc_small")
pbmc_small %>%  rename(s_score = nFeature_RNA)

#> # A Seurat-tibble abstraction: 80 x 15
#> # Features=230 | Active assay=RNA | Assays=RNA
#>    cell  orig.ident nCount_RNA s_score RNA_snn_res.0.8 letter.idents groups
#>    <chr> <fct>           <dbl>   <int> <fct>           <fct>         <chr> 
#>  1 ATGC… SeuratPro…         70      47 0               A             g2    
#>  2 CATG… SeuratPro…         85      52 0               A             g1    
#>  3 GAAC… SeuratPro…         87      50 1               B             g2    
#>  4 TGAC… SeuratPro…        127      56 0               A             g2    
#>  5 AGTC… SeuratPro…        173      53 0               A             g2    
#>  6 TCTG… SeuratPro…         70      48 0               A             g1    
#>  7 TGGT… SeuratPro…         64      36 0               A             g1    
#>  8 GCAG… SeuratPro…         72      45 0               A             g1    
#>  9 GATA… SeuratPro…         52      36 0               A             g1    
#> 10 AATG… SeuratPro…        100      41 0               A             g1    
#> # … with 70 more rows, and 8 more variables: RNA_snn_res.1 <fct>, PC_1 <dbl>,
#> #   PC_2 <dbl>, PC_3 <dbl>, PC_4 <dbl>, PC_5 <dbl>, tSNE_1 <dbl>, tSNE_2 <dbl>

`%>%` = magrittr::`%>%`

`%>%` = magrittr::`%>%`
data("pbmc_small")
pbmc_small %>% left_join(pbmc_small %>% distinct(groups) %>% mutate(new_column = 1:2))

#> tidyseurat says: A data frame is returned for independent data analysis.

#> Joining, by = "groups"

#> # A Seurat-tibble abstraction: 80 x 16
#> # Features=230 | Active assay=RNA | Assays=RNA
#>    cell  orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.8 letter.idents groups
#>    <chr> <fct>           <dbl>        <int> <fct>           <fct>         <chr> 
#>  1 ATGC… SeuratPro…         70           47 0               A             g2    
#>  2 CATG… SeuratPro…         85           52 0               A             g1    
#>  3 GAAC… SeuratPro…         87           50 1               B             g2    
#>  4 TGAC… SeuratPro…        127           56 0               A             g2    
#>  5 AGTC… SeuratPro…        173           53 0               A             g2    
#>  6 TCTG… SeuratPro…         70           48 0               A             g1    
#>  7 TGGT… SeuratPro…         64           36 0               A             g1    
#>  8 GCAG… SeuratPro…         72           45 0               A             g1    
#>  9 GATA… SeuratPro…         52           36 0               A             g1    
#> 10 AATG… SeuratPro…        100           41 0               A             g1    
#> # … with 70 more rows, and 9 more variables: RNA_snn_res.1 <fct>,
#> #   new_column <int>, PC_1 <dbl>, PC_2 <dbl>, PC_3 <dbl>, PC_4 <dbl>,
#> #   PC_5 <dbl>, tSNE_1 <dbl>, tSNE_2 <dbl>

`%>%` = magrittr::`%>%`

data("pbmc_small")
pbmc_small %>% inner_join(pbmc_small %>% distinct(groups) %>% mutate(new_column = 1:2) %>% slice(1))

#> tidyseurat says: A data frame is returned for independent data analysis.

#> Joining, by = "groups"

#> # A Seurat-tibble abstraction: 36 x 16
#> # Features=230 | Active assay=RNA | Assays=RNA
#>    cell  orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.8 letter.idents groups
#>    <chr> <fct>           <dbl>        <int> <fct>           <fct>         <chr> 
#>  1 ATGC… SeuratPro…         70           47 0               A             g2    
#>  2 GAAC… SeuratPro…         87           50 1               B             g2    
#>  3 TGAC… SeuratPro…        127           56 0               A             g2    
#>  4 AGTC… SeuratPro…        173           53 0               A             g2    
#>  5 AGGT… SeuratPro…         62           31 0               A             g2    
#>  6 GGGT… SeuratPro…        101           41 0               A             g2    
#>  7 CATG… SeuratPro…         51           26 0               A             g2    
#>  8 TACG… SeuratPro…         99           45 0               A             g2    
#>  9 GTAA… SeuratPro…         67           33 0               A             g2    
#> 10 TACA… SeuratPro…        109           41 0               A             g2    
#> # … with 26 more rows, and 9 more variables: RNA_snn_res.1 <fct>,
#> #   new_column <int>, PC_1 <dbl>, PC_2 <dbl>, PC_3 <dbl>, PC_4 <dbl>,
#> #   PC_5 <dbl>, tSNE_1 <dbl>, tSNE_2 <dbl>

`%>%` = magrittr::`%>%`

data("pbmc_small")
pbmc_small %>% right_join(pbmc_small %>% distinct(groups) %>% mutate(new_column = 1:2) %>% slice(1))

#> tidyseurat says: A data frame is returned for independent data analysis.

#> Joining, by = "groups"

#> # A Seurat-tibble abstraction: 36 x 16
#> # Features=230 | Active assay=RNA | Assays=RNA
#>    cell  orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.8 letter.idents groups
#>    <chr> <fct>           <dbl>        <int> <fct>           <fct>         <chr> 
#>  1 ATGC… SeuratPro…         70           47 0               A             g2    
#>  2 GAAC… SeuratPro…         87           50 1               B             g2    
#>  3 TGAC… SeuratPro…        127           56 0               A             g2    
#>  4 AGTC… SeuratPro…        173           53 0               A             g2    
#>  5 AGGT… SeuratPro…         62           31 0               A             g2    
#>  6 GGGT… SeuratPro…        101           41 0               A             g2    
#>  7 CATG… SeuratPro…         51           26 0               A             g2    
#>  8 TACG… SeuratPro…         99           45 0               A             g2    
#>  9 GTAA… SeuratPro…         67           33 0               A             g2    
#> 10 TACA… SeuratPro…        109           41 0               A             g2    
#> # … with 26 more rows, and 9 more variables: RNA_snn_res.1 <fct>,
#> #   new_column <int>, PC_1 <dbl>, PC_2 <dbl>, PC_3 <dbl>, PC_4 <dbl>,
#> #   PC_5 <dbl>, tSNE_1 <dbl>, tSNE_2 <dbl>

`%>%` = magrittr::`%>%`
data("pbmc_small")
pbmc_small %>% full_join(tibble::tibble(groups = "g1", other=1:4))

#> Joining, by = "groups"

#> tidyseurat says: This operation lead to duplicated cell names. A data frame is returned for independent data analysis.

#> # A tibble: 212 x 30
#>    cell  orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.8 letter.idents groups
#>    <chr> <fct>           <dbl>        <int> <fct>           <fct>         <chr> 
#>  1 ATGC… SeuratPro…         70           47 0               A             g2    
#>  2 CATG… SeuratPro…         85           52 0               A             g1    
#>  3 CATG… SeuratPro…         85           52 0               A             g1    
#>  4 CATG… SeuratPro…         85           52 0               A             g1    
#>  5 CATG… SeuratPro…         85           52 0               A             g1    
#>  6 GAAC… SeuratPro…         87           50 1               B             g2    
#>  7 TGAC… SeuratPro…        127           56 0               A             g2    
#>  8 AGTC… SeuratPro…        173           53 0               A             g2    
#>  9 TCTG… SeuratPro…         70           48 0               A             g1    
#> 10 TCTG… SeuratPro…         70           48 0               A             g1    
#> # … with 202 more rows, and 23 more variables: RNA_snn_res.1 <fct>, PC_1 <dbl>,
#> #   PC_2 <dbl>, PC_3 <dbl>, PC_4 <dbl>, PC_5 <dbl>, PC_6 <dbl>, PC_7 <dbl>,
#> #   PC_8 <dbl>, PC_9 <dbl>, PC_10 <dbl>, PC_11 <dbl>, PC_12 <dbl>, PC_13 <dbl>,
#> #   PC_14 <dbl>, PC_15 <dbl>, PC_16 <dbl>, PC_17 <dbl>, PC_18 <dbl>,
#> #   PC_19 <dbl>, tSNE_1 <dbl>, tSNE_2 <dbl>, other <int>

`%>%` = magrittr::`%>%`
data("pbmc_small")
pbmc_small %>%  slice(1)

#> # A Seurat-tibble abstraction: 1 x 15
#> # Features=230 | Active assay=RNA | Assays=RNA
#>   cell  orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.8 letter.idents groups
#>   <chr> <fct>           <dbl>        <int> <fct>           <fct>         <chr> 
#> 1 ATGC… SeuratPro…         70           47 0               A             g2    
#> # … with 8 more variables: RNA_snn_res.1 <fct>, PC_1 <dbl>, PC_2 <dbl>,
#> #   PC_3 <dbl>, PC_4 <dbl>, PC_5 <dbl>, tSNE_1 <dbl>, tSNE_2 <dbl>

`%>%` = magrittr::`%>%`
data("pbmc_small")
pbmc_small %>%  select(cell, orig.ident )

#> # A Seurat-tibble abstraction: 80 x 9
#> # Features=230 | Active assay=RNA | Assays=RNA
#>    cell          orig.ident      PC_1   PC_2   PC_3  PC_4   PC_5  tSNE_1  tSNE_2
#>    <chr>         <fct>          <dbl>  <dbl>  <dbl> <dbl>  <dbl>   <dbl>   <dbl>
#>  1 ATGCCAGAACGA… SeuratProje… -0.774  -0.900 -0.249 0.559  0.465   0.868  -8.10 
#>  2 CATGGCCTGTGC… SeuratProje… -0.0260 -0.347  0.665 0.418  0.585  -7.39   -8.77 
#>  3 GAACCTGATGAA… SeuratProje… -0.457   0.180  1.32  2.01  -0.482 -28.2     0.241
#>  4 TGACTGGATTCT… SeuratProje… -0.812  -1.38  -1.00  0.139 -1.60   16.3   -11.2  
#>  5 AGTCAGACTGCA… SeuratProje… -0.774  -0.900 -0.249 0.559  0.465   1.91  -11.2  
#>  6 TCTGATACACGT… SeuratProje… -0.774  -0.900 -0.249 0.559  0.465   3.15   -9.94 
#>  7 TGGTATCTAAAC… SeuratProje… -0.460  -1.19  -0.312 0.716 -1.65   17.9    -9.90 
#>  8 GCAGCTCTGTTT… SeuratProje… -0.900  -0.388  0.693 0.404  0.536  -6.49   -8.39 
#>  9 GATATAACACGC… SeuratProje… -0.774  -0.900 -0.249 0.559  0.465   1.33   -9.68 
#> 10 AATGTTGACAGT… SeuratProje… -0.488  -1.16  -0.306 0.702 -1.47   17.0    -9.43 
#> # … with 70 more rows

`%>%` = magrittr::`%>%`
data("pbmc_small")
pbmc_small %>%  sample_n(50)

#> # A Seurat-tibble abstraction: 50 x 15
#> # Features=230 | Active assay=RNA | Assays=RNA
#>    cell  orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.8 letter.idents groups
#>    <chr> <fct>           <dbl>        <int> <fct>           <fct>         <chr> 
#>  1 GATA… SeuratPro…        328           72 1               B             g1    
#>  2 GGCA… SeuratPro…        126           53 0               A             g1    
#>  3 ATGC… SeuratPro…         70           47 0               A             g2    
#>  4 AGAT… SeuratPro…        187           61 0               A             g2    
#>  5 TACA… SeuratPro…        108           44 0               A             g2    
#>  6 CATG… SeuratPro…         51           26 0               A             g2    
#>  7 GCAC… SeuratPro…        292           71 1               B             g2    
#>  8 CGTA… SeuratPro…        371           75 1               B             g1    
#>  9 TTAC… SeuratPro…        298           65 1               B             g1    
#> 10 ATAA… SeuratPro…         99           42 1               B             g2    
#> # … with 40 more rows, and 8 more variables: RNA_snn_res.1 <fct>, PC_1 <dbl>,
#> #   PC_2 <dbl>, PC_3 <dbl>, PC_4 <dbl>, PC_5 <dbl>, tSNE_1 <dbl>, tSNE_2 <dbl>

pbmc_small %>%  sample_frac(0.1)

#> # A Seurat-tibble abstraction: 8 x 15
#> # Features=230 | Active assay=RNA | Assays=RNA
#>   cell  orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.8 letter.idents groups
#>   <chr> <fct>           <dbl>        <int> <fct>           <fct>         <chr> 
#> 1 GATA… SeuratPro…        328           72 1               B             g1    
#> 2 GGCA… SeuratPro…        126           53 0               A             g1    
#> 3 ATGC… SeuratPro…         70           47 0               A             g2    
#> 4 AGAT… SeuratPro…        187           61 0               A             g2    
#> 5 TACA… SeuratPro…        108           44 0               A             g2    
#> 6 CATG… SeuratPro…         51           26 0               A             g2    
#> 7 GCAC… SeuratPro…        292           71 1               B             g2    
#> 8 CGTA… SeuratPro…        371           75 1               B             g1    
#> # … with 8 more variables: RNA_snn_res.1 <fct>, PC_1 <dbl>, PC_2 <dbl>,
#> #   PC_3 <dbl>, PC_4 <dbl>, PC_5 <dbl>, tSNE_1 <dbl>, tSNE_2 <dbl>

`%>%` = magrittr::`%>%`
data("pbmc_small")
pbmc_small %>%  pull(groups)

#> tidyseurat says: A data frame is returned for independent data analysis.

#>  [1] "g2" "g1" "g2" "g2" "g2" "g1" "g1" "g1" "g1" "g1" "g2" "g1" "g2" "g2" "g2"
#> [16] "g1" "g2" "g1" "g1" "g2" "g1" "g1" "g2" "g2" "g1" "g2" "g2" "g2" "g2" "g1"
#> [31] "g1" "g1" "g1" "g2" "g1" "g1" "g2" "g1" "g1" "g2" "g1" "g2" "g2" "g2" "g1"
#> [46] "g2" "g1" "g2" "g1" "g2" "g1" "g2" "g2" "g2" "g1" "g1" "g1" "g1" "g2" "g1"
#> [61] "g1" "g1" "g1" "g1" "g1" "g2" "g2" "g1" "g1" "g1" "g2" "g1" "g2" "g2" "g1"
#> [76] "g1" "g2" "g1" "g2" "g1"