Add differential transcription information to a tbl using edgeR.

join_transcripts() takes as imput a `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> | and returns a `tbl` with additional columns for the statistics from the hypothesis test.

join_transcripts(
  .data,
  transcripts = NULL,
  all = F,
  exclude_zeros = F,
  shape = "long"
)

Arguments

.data
A `tbl` formatted as | <SAMPLE> | <TRANSCRIPT> | <COUNT> | <...> |

transcripts
A formula with no response variable, representing the desired linear model

all
The name of the sample column

exclude_zeros
The name of the transcript/gene column

shape
The name of the transcript/gene abundance column

Value

A `tbl` with additional columns for the statistics from the hypothesis test (e.g., log fold change, p-value and false discovery rate).

Details

At the moment this function uses edgeR only, but other inference algorithms will be added in the near future.

Examples

# \donttest{

  join_transcripts(
      ~ condition,
      sample,
      transcript,
      `count`
  )

#> [1] "This function cannot be applied to this object"

# }