Extract and join information for transcripts.

join_transcripts() extracts and joins information for specified transcripts

join_transcripts(
  .data,
  transcripts = NULL,
  all = FALSE,
  exclude_zeros = FALSE,
  shape = "long"
)

Arguments

.data
A tidyseurat object

transcripts
A vector of transcript identifiers to join

all
If TRUE return all

exclude_zeros
If TRUE exclude zero values

shape
Format of the returned table "long" or "wide"

Value

A `tbl` containing the information.for the specified transcripts

Details

This function extracts information for specified transcripts and returns the information in either long or wide format.

Examples

pbmc_small %>% 
tidy %>% 
join_transcripts(transcripts = c("HLA-DRA", "LYZ"))

#> tidyseurat says: A data frame is returned for independent data analysis.

#> # A tibble: 160 x 18
#>    cell  transcript abundance_RNA orig.ident nCount_RNA nFeature_RNA
#>    <chr> <chr>              <dbl> <fct>           <dbl>        <int>
#>  1 ATGC… HLA-DRA             0    SeuratPro…         70           47
#>  2 ATGC… LYZ                 4.97 SeuratPro…         70           47
#>  3 CATG… HLA-DRA             4.78 SeuratPro…         85           52
#>  4 CATG… LYZ                 4.78 SeuratPro…         85           52
#>  5 GAAC… HLA-DRA             0    SeuratPro…         87           50
#>  6 GAAC… LYZ                 4.75 SeuratPro…         87           50
#>  7 TGAC… HLA-DRA             0    SeuratPro…        127           56
#>  8 TGAC… LYZ                 0    SeuratPro…        127           56
#>  9 AGTC… HLA-DRA             4.07 SeuratPro…        173           53
#> 10 AGTC… LYZ                 0    SeuratPro…        173           53
#> # … with 150 more rows, and 12 more variables: RNA_snn_res.0.8 <fct>,
#> #   letter.idents <fct>, groups <chr>, RNA_snn_res.1 <fct>, file <chr>,
#> #   PC_1 <dbl>, PC_2 <dbl>, PC_3 <dbl>, PC_4 <dbl>, PC_5 <dbl>, tSNE_1 <dbl>,
#> #   tSNE_2 <dbl>