Coerce lists, matrices, and more to data frames

`r lifecycle::badge("maturing")`

`as_tibble()` turns an existing object, such as a data frame or matrix, into a so-called tibble, a data frame with class [`tbl_df`]. This is in contrast with [tibble()], which builds a tibble from individual columns. `as_tibble()` is to [`tibble()`] as [base::as.data.frame()] is to [base::data.frame()].

`as_tibble()` is an S3 generic, with methods for: * [`data.frame`][base::data.frame()]: Thin wrapper around the `list` method that implements tibble's treatment of [rownames]. * [`matrix`][methods::matrix-class], [`poly`][stats::poly()], [`ts`][stats::ts()], [`table`][base::table()] * Default: Other inputs are first coerced with [base::as.data.frame()].

`r lifecycle::badge("maturing")`

`glimpse()` is like a transposed version of `print()`: columns run down the page, and data runs across. This makes it possible to see every column in a data frame. It's a little like [str()] applied to a data frame but it tries to show you as much data as possible. (And it always shows the underlying data, even when applied to a remote data source.)

This generic will be moved to pillar, and reexported from there as soon as it becomes available.

Arguments

rownames
How to treat existing row names of a data frame or matrix: * `NULL`: remove row names. This is the default. * `NA`: keep row names. * A string: the name of a new column. Existing rownames are transferred into this column and the `row.names` attribute is deleted. Read more in [rownames].

.name_repair
see tidyr

For compatibility only, do not use for new code.

x
An object to glimpse at.

width
Width of output: defaults to the setting of the option `tibble.width` (if finite) or the width of the console.

...
Unused, for extensibility.

Value

A tibble

x original x is (invisibly) returned, allowing `glimpse()` to be used within a data pipe line.

Row names

The default behavior is to silently remove row names.

New code should explicitly convert row names to a new column using the `rownames` argument.

For existing code that relies on the retention of row names, call `pkgconfig::set_config("tibble::rownames" = NA)` in your script or in your package's [.onLoad()] function.

Life cycle

Using `as_tibble()` for vectors is superseded as of version 3.0.0, prefer the more expressive maturing `as_tibble_row()` and `as_tibble_col()` variants for new code.

S3 methods

`glimpse` is an S3 generic with a customised method for `tbl`s and `data.frames`, and a default method that calls [str()].

See also

Examples

pbmc_small %>%  as_tibble()

#> # A tibble: 80 x 30
#>    cell  orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.8 letter.idents groups
#>    <chr> <fct>           <dbl>        <int> <fct>           <fct>         <chr> 
#>  1 ATGC… SeuratPro…         70           47 0               A             g2    
#>  2 CATG… SeuratPro…         85           52 0               A             g1    
#>  3 GAAC… SeuratPro…         87           50 1               B             g2    
#>  4 TGAC… SeuratPro…        127           56 0               A             g2    
#>  5 AGTC… SeuratPro…        173           53 0               A             g2    
#>  6 TCTG… SeuratPro…         70           48 0               A             g1    
#>  7 TGGT… SeuratPro…         64           36 0               A             g1    
#>  8 GCAG… SeuratPro…         72           45 0               A             g1    
#>  9 GATA… SeuratPro…         52           36 0               A             g1    
#> 10 AATG… SeuratPro…        100           41 0               A             g1    
#> # … with 70 more rows, and 23 more variables: RNA_snn_res.1 <fct>, file <chr>,
#> #   PC_1 <dbl>, PC_2 <dbl>, PC_3 <dbl>, PC_4 <dbl>, PC_5 <dbl>, PC_6 <dbl>,
#> #   PC_7 <dbl>, PC_8 <dbl>, PC_9 <dbl>, PC_10 <dbl>, PC_11 <dbl>, PC_12 <dbl>,
#> #   PC_13 <dbl>, PC_14 <dbl>, PC_15 <dbl>, PC_16 <dbl>, PC_17 <dbl>,
#> #   PC_18 <dbl>, PC_19 <dbl>, tSNE_1 <dbl>, tSNE_2 <dbl>

pbmc_small %>% tidy %>% glimpse()

#> Warning: `tidy()` was deprecated in tidyseurat 0.2.0.
#> tidyseurat says: tidy() is not needed anymore.

#> Formal class 'Seurat' [package "Seurat"] with 13 slots
#>   ..@ assays      :List of 1
#>   .. ..$ RNA:Formal class 'Assay' [package "Seurat"] with 8 slots
#>   ..@ meta.data   :'data.frame':	80 obs. of  8 variables:
#>   .. ..$ orig.ident     : Factor w/ 1 level "SeuratProject": 1 1 1 1 1 1 1 1 1 1 ...
#>   .. ..$ nCount_RNA     : num [1:80] 70 85 87 127 173 70 64 72 52 100 ...
#>   .. ..$ nFeature_RNA   : int [1:80] 47 52 50 56 53 48 36 45 36 41 ...
#>   .. ..$ RNA_snn_res.0.8: Factor w/ 2 levels "0","1": 1 1 2 1 1 1 1 1 1 1 ...
#>   .. ..$ letter.idents  : Factor w/ 2 levels "A","B": 1 1 2 1 1 1 1 1 1 1 ...
#>   .. ..$ groups         : chr [1:80] "g2" "g1" "g2" "g2" ...
#>   .. ..$ RNA_snn_res.1  : Factor w/ 3 levels "0","1","2": 1 1 1 1 1 1 1 1 1 1 ...
#>   .. ..$ file           : chr [1:80] "../data/sample2/outs/filtered_feature_bc_matrix/" "../data/sample1/outs/filtered_feature_bc_matrix/" "../data/sample2/outs/filtered_feature_bc_matrix/" "../data/sample2/outs/filtered_feature_bc_matrix/" ...
#>   ..@ active.assay: chr "RNA"
#>   ..@ active.ident: Factor w/ 3 levels "0","1","2": 1 1 1 1 1 1 1 1 1 1 ...
#>   .. ..- attr(*, "names")= chr [1:80] "ATGCCAGAACGACT" "CATGGCCTGTGCAT" "GAACCTGATGAACC" "TGACTGGATTCTCA" ...
#>   ..@ graphs      :List of 1
#>   .. ..$ RNA_snn:Formal class 'Graph' [package "Seurat"] with 7 slots
#>   ..@ neighbors   : list()
#>   ..@ reductions  :List of 2
#>   .. ..$ pca :Formal class 'DimReduc' [package "Seurat"] with 9 slots
#>   .. ..$ tsne:Formal class 'DimReduc' [package "Seurat"] with 9 slots
#>   ..@ images      : list()
#>   ..@ project.name: chr "SeuratProject"
#>   ..@ misc        : list()
#>   ..@ version     :Classes 'package_version', 'numeric_version'  hidden list of 1
#>   .. ..$ : int [1:3] 3 2 0
#>   ..@ commands    :List of 10
#>   .. ..$ NormalizeData.RNA       :Formal class 'SeuratCommand' [package "Seurat"] with 5 slots
#>   .. ..$ ScaleData.RNA           :Formal class 'SeuratCommand' [package "Seurat"] with 5 slots
#>   .. ..$ RunPCA.RNA              :Formal class 'SeuratCommand' [package "Seurat"] with 5 slots
#>   .. ..$ BuildSNN.RNA.pca        :Formal class 'SeuratCommand' [package "Seurat"] with 5 slots
#>   .. ..$ FindClusters            :Formal class 'SeuratCommand' [package "Seurat"] with 5 slots
#>   .. ..$ RunTSNE.pca             :Formal class 'SeuratCommand' [package "Seurat"] with 5 slots
#>   .. ..$ JackStraw.RNA.pca       :Formal class 'SeuratCommand' [package "Seurat"] with 5 slots
#>   .. ..$ ScoreJackStraw.pca      :Formal class 'SeuratCommand' [package "Seurat"] with 5 slots
#>   .. ..$ ProjectDim.RNA.pca      :Formal class 'SeuratCommand' [package "Seurat"] with 5 slots
#>   .. ..$ FindVariableFeatures.RNA:Formal class 'SeuratCommand' [package "Seurat"] with 5 slots
#>   ..@ tools       : list()