unnest

Given a regular expression with capturing groups, `extract()` turns each group into a new column. If the groups don't match, or the input is NA, the output will be NA.

Convenience function to paste together multiple columns into one.

Given either a regular expression or a vector of character positions, `separate()` turns a single character column into multiple columns.

Arguments

cols

<[`tidy-select`][tidyr_select]> Columns to unnest. If you `unnest()` multiple columns, parallel entries must be of compatible sizes, i.e. they're either equal or length 1 (following the standard tidyverse recycling rules).

names_sep

If `NULL`, the default, the names will be left as is. In `nest()`, inner names will come from the former outer names; in `unnest()`, the new outer names will come from the inner names.

If a string, the inner and outer names will be used together. In `nest()`, the names of the new outer columns will be formed by pasting together the outer and the inner column names, separated by `names_sep`. In `unnest()`, the new inner names will have the outer names (+ `names_sep`) automatically stripped. This makes `names_sep` roughly symmetric between nesting and unnesting.

keep_empty

See tidyr::unnest

names_repair

See tidyr::unnest

ptype

See tidyr::unnest

.drop

See tidyr::unnest

.id

tidyr::unnest

.sep

tidyr::unnest

.preserve

See tidyr::unnest

.data

A tbl. (See tidyr)

.names_sep

See tidyr::nest

into

Names of new variables to create as character vector. Use `NA` to omit the variable in the output.

regex

a regular expression used to extract the desired values. There should be one group (defined by `()`) for each element of `into`.

convert

If `TRUE`, will run [type.convert()] with `as.is = TRUE` on new columns. This is useful if the component columns are integer, numeric or logical.

NB: this will cause string `"NA"`s to be converted to `NA`s.

data

A data frame.

col

The name of the new column, as a string or symbol.

This argument is passed by expression and supports [quasiquotation][rlang::quasiquotation] (you can unquote strings and symbols). The name is captured from the expression with [rlang::ensym()] (note that this kind of interface where symbols do not represent actual objects is now discouraged in the tidyverse; we support it here for backward compatibility).

...

<[`tidy-select`][tidyr_select]> Columns to unite

na.rm

If `TRUE`, missing values will be remove prior to uniting each value.

remove

If `TRUE`, remove input columns from output data frame.

sep

Separator between columns.

If character, `sep` is interpreted as a regular expression. The default value is a regular expression that matches any sequence of non-alphanumeric values.

If numeric, `sep` is interpreted as character positions to split at. Positive values start at 1 at the far-left of the string; negative value start at -1 at the far-right of the string. The length of `sep` should be one less than `into`.

extra

If `sep` is a character vector, this controls what happens when there are too many pieces. There are three valid options:

* "warn" (the default): emit a warning and drop extra values. * "drop": drop any extra values without a warning. * "merge": only splits at most `length(into)` times

fill

If `sep` is a character vector, this controls what happens when there are not enough pieces. There are three valid options:

* "warn" (the default): emit a warning and fill from the right * "right": fill with missing values on the right * "left": fill with missing values on the left

Value

A Seurat object or a tibble depending on input

A Seurat object or a tibble depending on input

A Seurat object or a tibble depending on input

A Seurat object or a tibble depending on input

A Seurat object or a tibble depending on input

See also

Examples

library(dplyr)
data("pbmc_small")
pbmc_small %>%  nest(data = -groups) %>% unnest(data) 
#> # A Seurat-tibble abstraction: 80 × 8
#> # Features=230 | Cells=80 | Active assay=RNA | Assays=RNA
#>    .cell          orig.ident    nCount_…¹ nFeat…² RNA_s…³ lette…⁴ RNA_s…⁵ groups
#>    <chr>          <chr>             <dbl>   <int> <chr>   <chr>   <chr>   <chr> 
#>  1 ATGCCAGAACGACT SeuratProject        70      47 0       A       0       g2    
#>  2 GAACCTGATGAACC SeuratProject        87      50 1       B       0       g2    
#>  3 TGACTGGATTCTCA SeuratProject       127      56 0       A       0       g2    
#>  4 AGTCAGACTGCACA SeuratProject       173      53 0       A       0       g2    
#>  5 AGGTCATGAGTGTC SeuratProject        62      31 0       A       2       g2    
#>  6 GGGTAACTCTAGTG SeuratProject       101      41 0       A       2       g2    
#>  7 CATGAGACACGGGA SeuratProject        51      26 0       A       2       g2    
#>  8 TACGCCACTCCGAA SeuratProject        99      45 0       A       2       g2    
#>  9 GTAAGCACTCATTC SeuratProject        67      33 0       A       2       g2    
#> 10 TACATCACGCTAAC SeuratProject       109      41 0       A       2       g2    
#> # … with 70 more rows, and abbreviated variable names ¹​nCount_RNA,
#> #   ²​nFeature_RNA, ³​RNA_snn_res.0.8, ⁴​letter.idents, ⁵​RNA_snn_res.1

library(dplyr)
data("pbmc_small")
pbmc_small %>%  nest(data = -groups) %>% unnest(data) 
#> # A Seurat-tibble abstraction: 80 × 8
#> # Features=230 | Cells=80 | Active assay=RNA | Assays=RNA
#>    .cell          orig.ident    nCount_…¹ nFeat…² RNA_s…³ lette…⁴ RNA_s…⁵ groups
#>    <chr>          <chr>             <dbl>   <int> <chr>   <chr>   <chr>   <chr> 
#>  1 ATGCCAGAACGACT SeuratProject        70      47 0       A       0       g2    
#>  2 GAACCTGATGAACC SeuratProject        87      50 1       B       0       g2    
#>  3 TGACTGGATTCTCA SeuratProject       127      56 0       A       0       g2    
#>  4 AGTCAGACTGCACA SeuratProject       173      53 0       A       0       g2    
#>  5 AGGTCATGAGTGTC SeuratProject        62      31 0       A       2       g2    
#>  6 GGGTAACTCTAGTG SeuratProject       101      41 0       A       2       g2    
#>  7 CATGAGACACGGGA SeuratProject        51      26 0       A       2       g2    
#>  8 TACGCCACTCCGAA SeuratProject        99      45 0       A       2       g2    
#>  9 GTAAGCACTCATTC SeuratProject        67      33 0       A       2       g2    
#> 10 TACATCACGCTAAC SeuratProject       109      41 0       A       2       g2    
#> # … with 70 more rows, and abbreviated variable names ¹​nCount_RNA,
#> #   ²​nFeature_RNA, ³​RNA_snn_res.0.8, ⁴​letter.idents, ⁵​RNA_snn_res.1

data("pbmc_small")
pbmc_small %>%  extract(groups, into = "g", regex = "g([0-9])", convert = TRUE)
#> # A Seurat-tibble abstraction: 80 × 15
#> # Features=230 | Cells=80 | Active assay=RNA | Assays=RNA
#>    .cell    orig.…¹ nCoun…² nFeat…³ RNA_s…⁴ lette…⁵     g RNA_s…⁶    PC_1   PC_2
#>    <chr>    <fct>     <dbl>   <int> <fct>   <fct>   <int> <fct>     <dbl>  <dbl>
#>  1 ATGCCAG… Seurat…      70      47 0       A           2 0       -0.774  -0.900
#>  2 CATGGCC… Seurat…      85      52 0       A           1 0       -0.0260 -0.347
#>  3 GAACCTG… Seurat…      87      50 1       B           2 0       -0.457   0.180
#>  4 TGACTGG… Seurat…     127      56 0       A           2 0       -0.812  -1.38 
#>  5 AGTCAGA… Seurat…     173      53 0       A           2 0       -0.774  -0.900
#>  6 TCTGATA… Seurat…      70      48 0       A           1 0       -0.774  -0.900
#>  7 TGGTATC… Seurat…      64      36 0       A           1 0       -0.460  -1.19 
#>  8 GCAGCTC… Seurat…      72      45 0       A           1 0       -0.900  -0.388
#>  9 GATATAA… Seurat…      52      36 0       A           1 0       -0.774  -0.900
#> 10 AATGTTG… Seurat…     100      41 0       A           1 0       -0.488  -1.16 
#> # … with 70 more rows, 5 more variables: PC_3 <dbl>, PC_4 <dbl>, PC_5 <dbl>,
#> #   tSNE_1 <dbl>, tSNE_2 <dbl>, and abbreviated variable names ¹​orig.ident,
#> #   ²​nCount_RNA, ³​nFeature_RNA, ⁴​RNA_snn_res.0.8, ⁵​letter.idents,
#> #   ⁶​RNA_snn_res.1

data("pbmc_small")
pbmc_small %>%  unite("new_col", c(orig.ident, groups)) 
#> # A Seurat-tibble abstraction: 80 × 14
#> # Features=230 | Cells=80 | Active assay=RNA | Assays=RNA
#>    .cell   new_col nCoun…¹ nFeat…² RNA_s…³ lette…⁴ RNA_s…⁵    PC_1   PC_2   PC_3
#>    <chr>   <chr>     <dbl>   <int> <fct>   <fct>   <fct>     <dbl>  <dbl>  <dbl>
#>  1 ATGCCA… Seurat…      70      47 0       A       0       -0.774  -0.900 -0.249
#>  2 CATGGC… Seurat…      85      52 0       A       0       -0.0260 -0.347  0.665
#>  3 GAACCT… Seurat…      87      50 1       B       0       -0.457   0.180  1.32 
#>  4 TGACTG… Seurat…     127      56 0       A       0       -0.812  -1.38  -1.00 
#>  5 AGTCAG… Seurat…     173      53 0       A       0       -0.774  -0.900 -0.249
#>  6 TCTGAT… Seurat…      70      48 0       A       0       -0.774  -0.900 -0.249
#>  7 TGGTAT… Seurat…      64      36 0       A       0       -0.460  -1.19  -0.312
#>  8 GCAGCT… Seurat…      72      45 0       A       0       -0.900  -0.388  0.693
#>  9 GATATA… Seurat…      52      36 0       A       0       -0.774  -0.900 -0.249
#> 10 AATGTT… Seurat…     100      41 0       A       0       -0.488  -1.16  -0.306
#> # … with 70 more rows, 4 more variables: PC_4 <dbl>, PC_5 <dbl>, tSNE_1 <dbl>,
#> #   tSNE_2 <dbl>, and abbreviated variable names ¹​nCount_RNA, ²​nFeature_RNA,
#> #   ³​RNA_snn_res.0.8, ⁴​letter.idents, ⁵​RNA_snn_res.1

data("pbmc_small")
 un = pbmc_small %>%  unite("new_col", c(orig.ident, groups)) 
 un %>% separate(col = new_col, into= c("orig.ident", "groups"))
#> # A Seurat-tibble abstraction: 80 × 15
#> # Features=230 | Cells=80 | Active assay=RNA | Assays=RNA
#>    .cell   orig.…¹ groups nCoun…² nFeat…³ RNA_s…⁴ lette…⁵ RNA_s…⁶    PC_1   PC_2
#>    <chr>   <chr>   <chr>    <dbl>   <int> <fct>   <fct>   <fct>     <dbl>  <dbl>
#>  1 ATGCCA… Seurat… g2          70      47 0       A       0       -0.774  -0.900
#>  2 CATGGC… Seurat… g1          85      52 0       A       0       -0.0260 -0.347
#>  3 GAACCT… Seurat… g2          87      50 1       B       0       -0.457   0.180
#>  4 TGACTG… Seurat… g2         127      56 0       A       0       -0.812  -1.38 
#>  5 AGTCAG… Seurat… g2         173      53 0       A       0       -0.774  -0.900
#>  6 TCTGAT… Seurat… g1          70      48 0       A       0       -0.774  -0.900
#>  7 TGGTAT… Seurat… g1          64      36 0       A       0       -0.460  -1.19 
#>  8 GCAGCT… Seurat… g1          72      45 0       A       0       -0.900  -0.388
#>  9 GATATA… Seurat… g1          52      36 0       A       0       -0.774  -0.900
#> 10 AATGTT… Seurat… g1         100      41 0       A       0       -0.488  -1.16 
#> # … with 70 more rows, 5 more variables: PC_3 <dbl>, PC_4 <dbl>, PC_5 <dbl>,
#> #   tSNE_1 <dbl>, tSNE_2 <dbl>, and abbreviated variable names ¹​orig.ident,
#> #   ²​nCount_RNA, ³​nFeature_RNA, ⁴​RNA_snn_res.0.8, ⁵​letter.idents,
#> #   ⁶​RNA_snn_res.1