unnest

Given a regular expression with capturing groups, `extract()` turns each group into a new column. If the groups don't match, or the input is NA, the output will be NA.

Convenience function to paste together multiple columns into one.

Given either a regular expression or a vector of character positions, `separate()` turns a single character column into multiple columns.

Arguments

cols
<[`tidy-select`][tidyr_select]> Columns to unnest. If you `unnest()` multiple columns, parallel entries must be of compatible sizes, i.e. they're either equal or length 1 (following the standard tidyverse recycling rules).

names_sep
If `NULL`, the default, the names will be left as is. In `nest()`, inner names will come from the former outer names; in `unnest()`, the new outer names will come from the inner names.

If a string, the inner and outer names will be used together. In `nest()`, the names of the new outer columns will be formed by pasting together the outer and the inner column names, separated by `names_sep`. In `unnest()`, the new inner names will have the outer names (+ `names_sep`) automatically stripped. This makes `names_sep` roughly symmetric between nesting and unnesting.

keep_empty
See tidyr::unnest

names_repair
See tidyr::unnest

ptype
See tidyr::unnest

.drop
See tidyr::unnest

.id
tidyr::unnest

.preserve
See tidyr::unnest

.data
A tbl. (See tidyr)

.names_sep
See tidyr::nest

into
Names of new variables to create as character vector. Use `NA` to omit the variable in the output.

regex
a regular expression used to extract the desired values. There should be one group (defined by `()`) for each element of `into`.

convert
If `TRUE`, will run [type.convert()] with `as.is = TRUE` on new columns. This is useful if the component columns are integer, numeric or logical.

NB: this will cause string `"NA"`s to be converted to `NA`s.

data
A data frame.

col
The name of the new column, as a string or symbol.

This argument is passed by expression and supports [quasiquotation][rlang::quasiquotation] (you can unquote strings and symbols). The name is captured from the expression with [rlang::ensym()] (note that this kind of interface where symbols do not represent actual objects is now discouraged in the tidyverse; we support it here for backward compatibility).

...
<[`tidy-select`][tidyr_select]> Columns to unite

na.rm
If `TRUE`, missing values will be remove prior to uniting each value.

remove
If `TRUE`, remove input columns from output data frame.

sep
Separator between columns.

If character, `sep` is interpreted as a regular expression. The default value is a regular expression that matches any sequence of non-alphanumeric values.

If numeric, `sep` is interpreted as character positions to split at. Positive values start at 1 at the far-left of the string; negative value start at -1 at the far-right of the string. The length of `sep` should be one less than `into`.

extra
If `sep` is a character vector, this controls what happens when there are too many pieces. There are three valid options:

* "warn" (the default): emit a warning and drop extra values. * "drop": drop any extra values without a warning. * "merge": only splits at most `length(into)` times

fill
If `sep` is a character vector, this controls what happens when there are not enough pieces. There are three valid options:

* "warn" (the default): emit a warning and fill from the right * "right": fill with missing values on the right * "left": fill with missing values on the left

Value

A Seurat object or a tibble depending on input

A Seurat object or a tibble depending on input

A Seurat object or a tibble depending on input

A Seurat object or a tibble depending on input

A Seurat object or a tibble depending on input

See also

Examples

library(dplyr)
data("pbmc_small")
pbmc_small %>%  nest(data = -groups) %>% unnest(data) 

#> # A Seurat-tibble abstraction: 80 × 8
#> # Features=230 | Active assay=RNA | Assays=RNA
#>    cell  orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.8 letter.idents
#>    <chr> <chr>           <dbl>        <int> <chr>           <chr>        
#>  1 ATGC… SeuratPro…         70           47 0               A            
#>  2 GAAC… SeuratPro…         87           50 1               B            
#>  3 TGAC… SeuratPro…        127           56 0               A            
#>  4 AGTC… SeuratPro…        173           53 0               A            
#>  5 AGGT… SeuratPro…         62           31 0               A            
#>  6 GGGT… SeuratPro…        101           41 0               A            
#>  7 CATG… SeuratPro…         51           26 0               A            
#>  8 TACG… SeuratPro…         99           45 0               A            
#>  9 GTAA… SeuratPro…         67           33 0               A            
#> 10 TACA… SeuratPro…        109           41 0               A            
#> # … with 70 more rows, and 2 more variables: RNA_snn_res.1 <chr>, groups <chr>

library(dplyr)
data("pbmc_small")
pbmc_small %>%  nest(data = -groups) %>% unnest(data) 

#> # A Seurat-tibble abstraction: 80 × 8
#> # Features=230 | Active assay=RNA | Assays=RNA
#>    cell  orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.8 letter.idents
#>    <chr> <chr>           <dbl>        <int> <chr>           <chr>        
#>  1 ATGC… SeuratPro…         70           47 0               A            
#>  2 GAAC… SeuratPro…         87           50 1               B            
#>  3 TGAC… SeuratPro…        127           56 0               A            
#>  4 AGTC… SeuratPro…        173           53 0               A            
#>  5 AGGT… SeuratPro…         62           31 0               A            
#>  6 GGGT… SeuratPro…        101           41 0               A            
#>  7 CATG… SeuratPro…         51           26 0               A            
#>  8 TACG… SeuratPro…         99           45 0               A            
#>  9 GTAA… SeuratPro…         67           33 0               A            
#> 10 TACA… SeuratPro…        109           41 0               A            
#> # … with 70 more rows, and 2 more variables: RNA_snn_res.1 <chr>, groups <chr>

data("pbmc_small")
pbmc_small %>%  extract(groups, into = "g", regex = "g([0-9])", convert = TRUE)

#> # A Seurat-tibble abstraction: 80 × 15
#> # Features=230 | Active assay=RNA | Assays=RNA
#>    cell  orig.ident nCount_RNA nFeature_RNA RNA_snn_res.0.8 letter.idents     g
#>    <chr> <fct>           <dbl>        <int> <fct>           <fct>         <int>
#>  1 ATGC… SeuratPro…         70           47 0               A                 2
#>  2 CATG… SeuratPro…         85           52 0               A                 1
#>  3 GAAC… SeuratPro…         87           50 1               B                 2
#>  4 TGAC… SeuratPro…        127           56 0               A                 2
#>  5 AGTC… SeuratPro…        173           53 0               A                 2
#>  6 TCTG… SeuratPro…         70           48 0               A                 1
#>  7 TGGT… SeuratPro…         64           36 0               A                 1
#>  8 GCAG… SeuratPro…         72           45 0               A                 1
#>  9 GATA… SeuratPro…         52           36 0               A                 1
#> 10 AATG… SeuratPro…        100           41 0               A                 1
#> # … with 70 more rows, and 8 more variables: RNA_snn_res.1 <fct>, PC_1 <dbl>,
#> #   PC_2 <dbl>, PC_3 <dbl>, PC_4 <dbl>, PC_5 <dbl>, tSNE_1 <dbl>, tSNE_2 <dbl>

data("pbmc_small")
pbmc_small %>%  unite("new_col", c(orig.ident, groups)) 

#> # A Seurat-tibble abstraction: 80 × 14
#> # Features=230 | Active assay=RNA | Assays=RNA
#>    cell  new_col nCount_RNA nFeature_RNA RNA_snn_res.0.8 letter.idents
#>    <chr> <chr>        <dbl>        <int> <fct>           <fct>        
#>  1 ATGC… Seurat…         70           47 0               A            
#>  2 CATG… Seurat…         85           52 0               A            
#>  3 GAAC… Seurat…         87           50 1               B            
#>  4 TGAC… Seurat…        127           56 0               A            
#>  5 AGTC… Seurat…        173           53 0               A            
#>  6 TCTG… Seurat…         70           48 0               A            
#>  7 TGGT… Seurat…         64           36 0               A            
#>  8 GCAG… Seurat…         72           45 0               A            
#>  9 GATA… Seurat…         52           36 0               A            
#> 10 AATG… Seurat…        100           41 0               A            
#> # … with 70 more rows, and 8 more variables: RNA_snn_res.1 <fct>, PC_1 <dbl>,
#> #   PC_2 <dbl>, PC_3 <dbl>, PC_4 <dbl>, PC_5 <dbl>, tSNE_1 <dbl>, tSNE_2 <dbl>

data("pbmc_small")
 un = pbmc_small %>%  unite("new_col", c(orig.ident, groups)) 
 un %>% separate(col = new_col, into= c("orig.ident", "groups"))

#> # A Seurat-tibble abstraction: 80 × 15
#> # Features=230 | Active assay=RNA | Assays=RNA
#>    cell  orig.ident groups nCount_RNA nFeature_RNA RNA_snn_res.0.8 letter.idents
#>    <chr> <chr>      <chr>       <dbl>        <int> <fct>           <fct>        
#>  1 ATGC… SeuratPro… g2             70           47 0               A            
#>  2 CATG… SeuratPro… g1             85           52 0               A            
#>  3 GAAC… SeuratPro… g2             87           50 1               B            
#>  4 TGAC… SeuratPro… g2            127           56 0               A            
#>  5 AGTC… SeuratPro… g2            173           53 0               A            
#>  6 TCTG… SeuratPro… g1             70           48 0               A            
#>  7 TGGT… SeuratPro… g1             64           36 0               A            
#>  8 GCAG… SeuratPro… g1             72           45 0               A            
#>  9 GATA… SeuratPro… g1             52           36 0               A            
#> 10 AATG… SeuratPro… g1            100           41 0               A            
#> # … with 70 more rows, and 8 more variables: RNA_snn_res.1 <fct>, PC_1 <dbl>,
#> #   PC_2 <dbl>, PC_3 <dbl>, PC_4 <dbl>, PC_5 <dbl>, tSNE_1 <dbl>, tSNE_2 <dbl>